+++ tests/test_connectors.py
@@ -4,7 +4,6 @@ import tempfile
 import unittest
 
 import ayeaye
-from ayeaye.connectors.flowerpot import FlowerpotEngine, FlowerPotConnector
 from ayeaye.connectors.json_connector import JsonConnector
 
 PROJECT_TEST_PATH = os.path.dirname(os.path.abspath(__file__))
@@ -14,43 +13,6 @@ EXAMPLE_JSON_PATH = os.path.join(PROJECT_TEST_PATH, "data", "london_weather.json
 
 
 class TestConnectors(unittest.TestCase):
-    def test_flowerpot_deserialize(self):
-        test_string = bytes(
-            '{"availability": "apple", "referential": "raspberry"}\n{"availability": "anchor", "referential": "rudder"}',
-            encoding="utf-8",
-        )
-        result = FlowerpotEngine._deserialize_ndjson_string(test_string)
-        assert result[0] == {"availability": "apple", "referential": "raspberry"}
-        assert result[1] == {"availability": "anchor", "referential": "rudder"}
-
-    def test_iterate_over_json_lines(self):
-        with tarfile.open(EXAMPLE_FLOWERPOT_PATH, "r:gz") as tf:
-            reader = FlowerpotEngine(tf)
-            results = list(reader.items())
-            assert len(results) == 4
-            assert "availability" in results[0]
-            assert "referential" in results[0]
-
-    def test_flowerpot_all_items(self):
-        """
-        Iterate all the data items in all the files in the example flowerpot.
-        """
-        c = FlowerPotConnector(engine_url="flowerpot://" + EXAMPLE_FLOWERPOT_PATH)
-        all_items = [(r.availability, r.referential) for r in c]
-        all_items.sort()
-        expected = "[('acoustic', 'rap'), ('anchor', 'rudder'), ('antenna', 'receive'), ('apple', 'raspberry')]"
-        assert expected == str(all_items)
-
-    def test_flowerpot_query_one_file(self):
-        """
-        The 'table' kwarg gets rows from all files that start with that string.
-        """
-        c = FlowerPotConnector(engine_url="flowerpot://" + EXAMPLE_FLOWERPOT_PATH)
-        some_items = [(r.availability, r.referential) for r in c.query(table="test_a")]
-        some_items.sort()
-        expected = "[('anchor', 'rudder'), ('apple', 'raspberry')]"
-        assert expected == str(some_items)
-
     def test_json_basics(self):
         c = JsonConnector(engine_url="json://" + EXAMPLE_JSON_PATH)
         self.assertEqual("London", c.data.name)