+++ tests/test_model_collection.py
@@ -230,6 +230,7 @@ class TestModelCollection(unittest.TestCase):
         msg = "There are 3 models"
         self.assertEqual(3, len(models), msg)
 
+    @unittest.skip("TODO: Incomplete provenance code")
     def test_data_provenance_multiple_graphs(self):
         """
         The set of models contains two separate (none-connected) graphs.